Version 2.4.0
VDR version 2.4.0 is now available at

      ftp://ftp.tvdr.de/vdr/vdr-2.4.0.tar.bz2

A 'diff' against the previous version is available at

      ftp://ftp.tvdr.de/vdr/Developer/vdr-2.3.9-2.4.0.diff

MD5 checksums:

12c6a3abeadfa915fcfe736bb047a3ab  vdr-2.4.0.tar.bz2
6d98c2d619e82f3d4503146602fccde7  vdr-2.3.9-2.4.0.diff

A summary of all the major changes since the last stable version 2.2.0 can be found at

     http://www.tvdr.de/changelog.htm

When updating from an earlier version of VDR please make sure you read the INSTALL
and MANUAL files that come with the VDR source _before_ doing so!
Please make sure you have backup copies of all your configuration files,
and verify carefully that your timers will be set to the correct channels
after switching to this new version.

Thanks to the many people who have contributed in the making, testing and
debugging of this new version of VDR, and also to all users who just
enjoy VDR!