vdr-2.4.0-08-add-multi-frontend-support.diff
Added support for DVB devices with more than one frontend that all use the same dvr and demux. Note that in order for this to work, you must not set symbolic links like "demux1 -> demux0" and "dvr1 -> dvr0", as is mentioned in some user manuals of multi frontend DVB cards.